pyproject.toml
[project]
name = "wb-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-htmx>=1.23.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular[sidecar]>=0.28.0",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.1",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"
